requirements.txt
Django==4.1.7
django-filter==22.1
django-mssql-backend==2.8.1
djangorestframework==3.14.0
djangorestframework-csv==2.1.1
djangorestframework-simplejwt==5.2.2
drf-extra-fields==3.4.1
mssql-django==1.2
psycopg2==2.9.5
PyJWT==2.6.0
pymssql==2.2.7
pyodbc==4.0.35
python-dotenv==1.0.0
sql-server.pyodbc==1.0
djoser==2.2.0
drf-nested-routers==0.93.4
drf-yasg==1.21.5
requests==2.31.0
gunicorn==20.1.0
Pillow==9.5.0